Update the doc example
The `assert`s at the end appear to be written with `assert_eq` in mind, but with the `assert` macro, the second argument is a message parameter to be displayed on panics. Since these are testing `bool`s, we can leave them as `assert`s and remove the second parameters.

Another object, `d`, was added to include an object being removed from a `Heap`.